[ 🔧 | #47 | Refactor ] auth 리팩토링
