+++ src/pages/auth/LoginCompletePage/LoginCompletePage.jsx
@@ -1,0 +1,3 @@
+export default function LoginCompletePage() {
+	return <div>로그인 완료 페이지</div>;
+}
