--- src/pages/auth/ProfileSettingPage/index.jsx
@@ -1,3 +1,0 @@
-export default function ProfileSettingPage() {
-  return <div>profile setting</div>;
-}
