+++ src/pages/auth/SchoolSettingPage/components/input.jsx
